# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Video Analysis Pipeline",
    "version": "0.1.0"
  },
  "paths": {
    "/projects": {
      "get": {
        "summary": "List Projects",
        "description": "分析済みプロジェクトの一覧を取得.",
        "operationId": "list_projects_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectSummary"
                  },
                  "type": "array",
                  "title": "Response List Projects Projects Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Project",
        "description": "動画ファイルを受け取りプロジェクトを新規作成する.",
        "operationId": "create_project_projects_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_project_projects_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectCreatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/media": {
      "get": {
        "summary": "Get Project Media",
        "description": "プロジェクトの元メディアを返却.",
        "operationId": "get_project_media_projects__project_id__media_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/analyze": {
      "post": {
        "summary": "Start Analysis",
        "description": "分析パイプラインをバックグラウンドで起動する.",
        "operationId": "start_analysis_projects__project_id__analyze_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Start Analysis Projects  Project Id  Analyze Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/analysis-status": {
      "get": {
        "summary": "Get Analysis Status",
        "description": "進行状況と中間結果を返却する.",
        "operationId": "get_analysis_status_projects__project_id__analysis_status_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/report": {
      "get": {
        "summary": "Get Final Report",
        "description": "最終レポートを取得する.",
        "operationId": "get_final_report_projects__project_id__report_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_project_projects_post": {
        "properties": {
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "default"
          },
          "video_file": {
            "type": "string",
            "format": "binary",
            "title": "Video File"
          }
        },
        "type": "object",
        "required": [
          "company_name",
          "product_name",
          "title",
          "video_file"
        ],
        "title": "Body_create_project_projects_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectCreatedResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type"
          },
          "media_url": {
            "type": "string",
            "title": "Media Url"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "analysis_progress": {
            "type": "number",
            "title": "Analysis Progress"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "プロジェクト作成日時"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_name",
          "product_name",
          "title",
          "model",
          "file_name",
          "media_type",
          "media_url",
          "status",
          "analysis_progress",
          "created_at"
        ],
        "title": "ProjectCreatedResponse",
        "description": "POST /projects のレスポンス."
      },
      "ProjectReportResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type"
          },
          "media_url": {
            "type": "string",
            "title": "Media Url"
          },
          "final_report": {
            "$ref": "#/components/schemas/FinalReport"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_name",
          "product_name",
          "title",
          "model",
          "media_type",
          "media_url",
          "final_report"
        ],
        "title": "ProjectReportResponse"
      },
      "ProjectStatusResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type"
          },
          "media_url": {
            "type": "string",
            "title": "Media Url"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "analysis_progress": {
            "type": "number",
            "title": "Analysis Progress"
          },
          "analysis_started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Started At"
          },
          "analysis_completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Completed At"
          },
          "analysis_duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Duration Seconds"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/AnalysisStep"
            },
            "type": "array",
            "title": "Steps"
          },
          "logs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Logs"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_name",
          "product_name",
          "title",
          "model",
          "media_type",
          "media_url",
          "status",
          "analysis_progress",
          "steps",
          "logs"
        ],
        "title": "ProjectStatusResponse"
      },
      "ProjectSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type"
          },
          "media_url": {
            "type": "string",
            "title": "Media Url"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "analysis_progress": {
            "type": "number",
            "title": "Analysis Progress"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_name",
          "product_name",
          "title",
          "model",
          "media_type",
          "media_url",
          "status",
          "analysis_progress",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}